+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/KtLint.kt
@@ -2,43 +2,25 @@ package com.pinterest.ktlint.core
 
 import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.internal.EditorConfigLoader
+import com.pinterest.ktlint.core.internal.initPsiFileFactory
 import java.nio.file.FileSystems
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.util.ArrayList
 import java.util.HashSet
-import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
-import org.jetbrains.kotlin.cli.common.messages.MessageCollector
-import org.jetbrains.kotlin.cli.jvm.compiler.EnvironmentConfigFiles
-import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
-import org.jetbrains.kotlin.com.intellij.mock.MockProject
-import org.jetbrains.kotlin.com.intellij.openapi.Disposable
-import org.jetbrains.kotlin.com.intellij.openapi.diagnostic.DefaultLogger
-import org.jetbrains.kotlin.com.intellij.openapi.diagnostic.Logger as DiagnosticLogger
-import org.jetbrains.kotlin.com.intellij.openapi.extensions.ExtensionPoint
-import org.jetbrains.kotlin.com.intellij.openapi.extensions.Extensions.getRootArea
 import org.jetbrains.kotlin.com.intellij.openapi.util.Key
-import org.jetbrains.kotlin.com.intellij.openapi.util.UserDataHolderBase
-import org.jetbrains.kotlin.com.intellij.pom.PomModel
-import org.jetbrains.kotlin.com.intellij.pom.PomModelAspect
-import org.jetbrains.kotlin.com.intellij.pom.PomTransaction
-import org.jetbrains.kotlin.com.intellij.pom.impl.PomTransactionBase
-import org.jetbrains.kotlin.com.intellij.pom.tree.TreeAspect
 import org.jetbrains.kotlin.com.intellij.psi.PsiComment
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiErrorElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiFileFactory
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
-import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.TreeCopyHandler
-import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.psi.KtAnnotated
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.psiUtil.endOffset
 import org.jetbrains.kotlin.psi.psiUtil.startOffset
-import sun.reflect.ReflectionFactory
 
 object KtLint {
 
@@ -49,10 +31,10 @@ object KtLint {
     private const val UTF8_BOM = "\uFEFF"
     const val STDIN_FILE = "<stdin>"
 
-    private val psiFileFactory: PsiFileFactory
-    private val nullSuppression = { _: Int, _: String, _: Boolean -> false }
-
+    private val psiFileFactory: PsiFileFactory = initPsiFileFactory()
     private val editorConfigLoader = EditorConfigLoader(FileSystems.getDefault())
+
+    private val nullSuppression = { _: Int, _: String, _: Boolean -> false }
 
     /**
      * @param fileName path of file to lint/format
@@ -81,57 +63,6 @@ object KtLint {
         }
 
         internal val isStdIn: Boolean get() = fileName == STDIN_FILE
-    }
-
-    init {
-        // do not print anything to the stderr when lexer is unable to match input
-        class LoggerFactory : DiagnosticLogger.Factory {
-            override fun getLoggerInstance(p: String): DiagnosticLogger = object : DefaultLogger(null) {
-                override fun warn(message: String?, t: Throwable?) {}
-                override fun error(message: String?, vararg details: String?) {}
-            }
-        }
-        DiagnosticLogger.setFactory(LoggerFactory::class.java)
-        val compilerConfiguration = CompilerConfiguration()
-        compilerConfiguration.put(CLIConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
-        val project = KotlinCoreEnvironment.createForProduction(
-            Disposable {},
-            compilerConfiguration,
-            EnvironmentConfigFiles.JVM_CONFIG_FILES
-        ).project
-        // everything below (up to PsiFileFactory.getInstance(...)) is to get AST mutations (`ktlint -F ...`) working
-        // otherwise it's not needed
-        val pomModel: PomModel = object : UserDataHolderBase(), PomModel {
-
-            override fun runTransaction(transaction: PomTransaction) {
-                (transaction as PomTransactionBase).run()
-            }
-
-            @Suppress("UNCHECKED_CAST")
-            override fun <T : PomModelAspect> getModelAspect(aspect: Class<T>): T? {
-                if (aspect == TreeAspect::class.java) {
-                    // using approach described in https://git.io/vKQTo due to the magical bytecode of TreeAspect
-                    // (check constructor signature and compare it to the source)
-                    // (org.jetbrains.kotlin:kotlin-compiler-embeddable:1.0.3)
-                    val constructor = ReflectionFactory.getReflectionFactory().newConstructorForSerialization(
-                        aspect,
-                        Any::class.java.getDeclaredConstructor(*arrayOfNulls<Class<*>>(0))
-                    )
-                    return constructor.newInstance(*emptyArray()) as T
-                }
-                return null
-            }
-        }
-        val extensionPoint = "org.jetbrains.kotlin.com.intellij.treeCopyHandler"
-        val extensionClassName = TreeCopyHandler::class.java.name!!
-        for (area in arrayOf(project.extensionArea, getRootArea())) {
-            if (!area.hasExtensionPoint(extensionPoint)) {
-                area.registerExtensionPoint(extensionPoint, extensionClassName, ExtensionPoint.Kind.INTERFACE)
-            }
-        }
-        project as MockProject
-        project.registerService(PomModel::class.java, pomModel)
-        psiFileFactory = PsiFileFactory.getInstance(project)
     }
 
     /**